don't copy premium addons to the mirrors (bug 679132)

+++ apps/files/models.py
@@ -216,6 +216,8 @@ class File(amo.models.OnChangeMixin, amo.models.ModelBase):
 
     @property
     def mirror_file_path(self):
+        if self.version.addon.is_premium():
+            return
         return os.path.join(settings.MIRROR_STAGE_PATH,
                             str(self.version.addon_id), self.filename)
 
@@ -253,7 +255,8 @@ class File(amo.models.OnChangeMixin, amo.models.ModelBase):
         src, dst = self.file_path, self.guarded_file_path
         self.mv(src, dst, 'Moving disabled file: %s => %s')
         # Remove the file from the mirrors if necessary.
-        if os.path.exists(smart_str(self.mirror_file_path)):
+        if (self.mirror_file_path and
+            os.path.exists(smart_str(self.mirror_file_path))):
             log.info('Unmirroring disabled file: %s'
                      % self.mirror_file_path)
             os.remove(self.mirror_file_path)
@@ -270,6 +273,9 @@ class File(amo.models.OnChangeMixin, amo.models.ModelBase):
         try:
             if os.path.exists(self.file_path):
                 dst = self.mirror_file_path
+                if not dst:
+                    return
+
                 log.info('Moving file to mirror: %s => %s'
                          % (self.file_path, dst))
                 if not os.path.exists(os.path.dirname(dst)):
